Move appearance and general settings all into 1 directive
