+++ app/common/locales/en.json
@@ -358,12 +358,15 @@
     },
     "settings": {
         "appearance" : "Appearance Settings",
-        "appearance_header_image" : "Header Image",
+        "appearance_header_image" : "Custom Background Image",
+        "appearance_header_image_instructions" : "Image needs to be >1000px wide.",
+        "current_header" : "Current background image",
         "site" : "Site Settings",
         "site_name" : "Site name",
         "site_email" : "Contact email address",
         "site_email_note" : "Use caution, as this will be visible to anyone on your site.",
         "site_description" : "Describe your site",
+        "site_description_placeholder" : "An optional one sentence description",
         "site_timezone" : "Site timezone",
         "site_language" : "Site language",
         "save" : "Save",
