--- app/setting/controllers/setting-appearance-controller.js
@@ -1,14 +1,0 @@
-module.exports = [
-    '$scope',
-    '$translate',
-    'ConfigEndpoint',
-function (
-    $scope,
-    $translate,
-    ConfigEndpoint
-) {
-    $translate('tool.manage_appearance').then(function (title) {
-        $scope.title = title;
-        $scope.$emit('setPageTitle', title);
-    });
-}];
