+++ app/setting/setting-module.js
@@ -8,6 +8,6 @@ angular.module('ushahidi.tools', [
 .directive('formEditor', require('./directives/setting-form-editor-directive.js'))
 .directive('formStageEditor', require('./directives/setting-form-stage-editor-directive.js'))
 
-.directive('appearanceEditor', require('./directives/setting-appearance-directive.js'))
+.directive('settingsEditor', require('./directives/setting-editor-directive.js'))
 
 .config(require('./setting-routes.js'));
