+++ server/www/templates/partials/main-menu.html
@@ -17,7 +17,6 @@
             </div>
             <ul class="toggle-content" dropdown-menu>
                 <li><a ng-href="/settings/general" translate>nav.general</a></li>
-                <li><a ng-href="/settings/appearance" translate>nav.appearance</a></li>
                 <li><a ng-href="/settings/map-settings" translate>nav.map_settings</a></li>
                 <li><a ng-href="/settings/datasources" translate>nav.data_sources</a></li>
                 <!-- <li><a ng-href="/settings/plugins" translate>nav.plugins</a></li>-->
