--- server/www/templates/settings/appearance-editor.html
@@ -1,14 +1,0 @@
- <form id="appearance">
-
-    <label class="input-label" for="appearance-header-image" translate>settings.appearance_header_image</label>
-     <input id="appearance-header-image" name="file" type="file" onchange="angular.element(this).scope().fileNameChanged(this)" />
-      <div class="divider white"></div>
-
-      <div class="save-button">
-         <button class="save" ng-click="updateConfig('site', appearance)"
-                    ng-disabled="form.$invalid || isUnchanged(appearance) || saving_config['appearance']">
-              <i class="fa fa-cog fa-spin" ng-show="saving_config['site']"></i>
-              <span translate>settings.save</span>
-         </button>
-     </div>
-</form>
