--- server/www/templates/settings/appearance.html
@@ -1,11 +1,0 @@
-<main class="page-content has-aside" role="main">
-    <ul class="breadcrumb">
-      <li><a ng-href="/settings" translate>nav.settings</a></li>
-      <li class="active" translate>tool.appearance_settings</li>
-    </ul>
-    <div class="main-col">
-        <h1 translate>settings.appearance</h1>
-        <div class="divider white"></div>
-            <appearance-editor></appearance-editor>
-        </div>
-</main>
