+++ server/www/templates/settings/settings-editor.html
@@ -1,0 +1,37 @@
+ <form id="site">
+    <label class="input-label" for="site-settings-name" translate>settings.site_name</label>
+    <input id="site-settings-name" type="text" ng-model="site.name" required>
+
+    <label class="input-label" for="site-settings-description" translate>settings.site_description</label>
+    <textarea id="site-settings-description" type="text"
+            ng-model="site.description" placeholder="{{'settings.site_description_placeholder' | translate}}"></textarea>
+
+    <label class="input-label" for="appearance-header-image" translate>settings.appearance_header_image</label>
+        <p class="small" translate>settings.appearance_header_image_instructions</p>
+    <input id="appearance-header-image" name="file" type="file" onchange="angular.element(this).scope().fileNameChanged(this)" />
+    <p class="small" ng-show="site.image_header"><a href="{{ site.image_header }}" translate>settings.current_header</a> <button type="button" ng-click="clearHeader()" class="button button-secondary icon-only trash"></button></p>
+
+    <label class="input-label" for="site-settings-email" translate>settings.site_email</label>
+    <p class="small" translate>settings.site_email_note</p>
+    <input id="site-settings-email" type="email" ng-model="site.email">
+
+    <label class="input-label" for="site-settings-timezone" translate>settings.site_timezone</label>
+    <div class="custom-select">
+        <select id="site-settings-timezone" ng-model="site.timezone" ng-options="timezone for timezone in timezones | orderBy | unique"></select>
+    </div>
+
+    <label class="input-label" for="site-settings-language" translate>settings.site_language</label>
+    <div class="custom-select">
+        <select id="site-settings-language" ng-model="site.language" ng-options="lang.code as 'languages.' + lang.code  | translate for lang in languages"></select>
+    </div>
+
+    <div class="divider white"></div>
+
+    <div class="save-button">
+        <button class="save" ng-click="updateConfig()"
+            ng-disabled="form.$invalid || isUnchanged(site) || saving_config">
+                <i class="fa fa-cog fa-spin" ng-show="saving_config"></i>
+                <span translate>settings.save</span>
+        </button>
+    </div>
+</form>
