+++ server/www/templates/settings/settings.html
@@ -6,40 +6,6 @@
     <div class="main-col">
         <h1 translate>settings.site</h1>
         <div class="divider white"></div>
-        <form id="site">
-
-            <label class="input-label" for="site-settings-name" translate>settings.site_name</label>
-            <input id="site-settings-name" type="text"
-                    ng-model="site.name">
-
-            <label class="input-label" for="site-settings-description" translate>settings.site_description</label>
-            <input id="site-settings-description" type="text"
-                    ng-model="site.description">
-
-            <label class="input-label" for="site-settings-email" translate>settings.site_email</label>
-            <p class="small" translate>settings.site_email_note</p>
-            <input id="site-settings-email" type="email"
-                    ng-model="site.email">
-
-            <label class="input-label" for="site-settings-timezone" translate>settings.site_timezone</label>
-            <div class="custom-select">
-                <select id="site-settings-timezone" ng-model="site.timezone" ng-options="timezone for timezone in timezones | orderBy | unique"></select>
-            </div>
-
-            <label class="input-label" for="site-settings-language" translate>settings.site_language</label>
-            <div class="custom-select">
-                <select id="site-settings-language" ng-model="site.language" ng-options="lang.code as 'languages.' + lang.code  | translate for lang in languages"></select>
-            </div>
-
-            <div class="divider white"></div>
-
-            <div class="save-button">
-                <button class="save" ng-click="updateConfig('site', site)"
-                    ng-disabled="form.$invalid || isUnchanged(site) || saving_config['site']">
-                        <i class="fa fa-cog fa-spin" ng-show="saving_config['site']"></i>
-                        <span translate>settings.save</span>
-                </button>
-            </div>
-        </form>
+        <settings-editor></settings-editor>
     </div>
 </main>
